+++ tests/Feature/Commands/CacheCommandTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace SebastiaanLuca\Module\Tests\Feature\Commands;
 
-use Illuminate\Filesystem\Filesystem;
 use SebastiaanLuca\Module\ModuleServiceProvider;
 use SebastiaanLuca\Module\Tests\TestCase;
 
@@ -36,17 +35,5 @@ class CacheCommandTest extends TestCase
         return [
             ModuleServiceProvider::class,
         ];
-    }
-
-    /**
-     * Setup the test environment.
-     *
-     * @return void
-     */
-    protected function setUp() : void
-    {
-        parent::setUp();
-
-        app(Filesystem::class)->copyDirectory(__DIR__ . '/../../resources/modules/', base_path('modules/'));
     }
 }