+++ CHANGELOG.md
@@ -15,11 +15,16 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 - Renamed `path` config entry to `directories`
 - Only autoload test paths in development or test environments
+- Refresh modules after creating a new one
 
 ### Fixed
 
 - Ignore module paths that don't exist
 - Providers not getting registered
+- Register autoloading command not correctly using config paths
+- Check for valid composer.json entries before merging config
+- Don't write composer.json config when there was none in the first place
+- Show error if module already exists when creating it
 
 ## 0.2.10 (2018-08-06)